// src/types/anime-news.ts
export interface AnimeNews {
    pagination: Pagination;
    data: AnimeNewsArticleEntry[];
}

export interface AnimeNewsArticleEntry {
    mal_id: number;
    url: string;
    title: string;
    date: string;
    author_username: string;
    author_url: string;
    forum_url: string;
    images: Images;
    comments: number;
    excerpt: string;
}

export interface Images {
    jpg: Jpg;
}

export interface Jpg {
    image_url: string;
}

export interface Pagination {
    last_visible_page: number;
    has_next_page: boolean;
}